pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "croak-cv"
version = "1.0.0-alpha"
description = "Computer Recognition Orchestration Agent Kit - Agentic framework for object detection"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "vfrog.ai", email = "hello@vfrog.ai" }
]
keywords = [
    "computer-vision",
    "object-detection",
    "machine-learning",
    "yolo",
    "ai-agents",
    "mlops"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "mlflow>=2.0.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
modal = ["modal>=0.50.0"]
wandb = ["wandb>=0.15.0"]
tensorrt = ["tensorrt>=8.6.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
all = [
    "croak-cv[modal,wandb,tensorrt,dev]"
]

[project.scripts]
croak = "croak.cli:main"

[project.urls]
Homepage = "https://github.com/vfrog-ai/croak"
Documentation = "https://croak.vfrog.ai"
Repository = "https://github.com/vfrog-ai/croak"
Issues = "https://github.com/vfrog-ai/croak/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/croak"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=croak --cov-report=term-missing"
